Review feedback : remove burn token incorrect script from all references in the code.. return more DOS status from CheckColorIdentifierValidity and remove invalid-colorid returned from its return code. update documentation and move all vulnerability test scenarios to feature_coloredcoin
