+++ src/txdb.cpp
@@ -453,5 +453,5 @@ bool CBlockTreeDB::LoadIssuedColorIds(std::set<ColorIdentifier>& colorIds)
         colorIds.insert(ColorIdentifier(key.second));
         pcursor->Next();
     }
-    return true;
+    return !pcursor->GetError();
 }
