+++ src/wallet/rpcwallet.cpp
@@ -4506,7 +4506,7 @@ UniValue transfertoken(const JSONRPCRequest& request)
 static CTransactionRef BurnToken(CWallet * const pwallet, const ColorIdentifier& colorId, CAmount nValue)
 {
     //this is to make create transaction recognize this as colored transaction
-    CScript scriptPubKey = CScript() << colorId.toVector() << OP_COLOR << OP_TRUE;
+    CScript scriptPubKey = CScript() << colorId.toVector() << OP_COLOR << OP_FALSE;
     mapValue_t mapValue;
     mapValue["comment"] = colorId.toHexString();
